requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.0.8
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.5
Pygments==2.6.1
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
